Implementato Security.java
create salt and get hash
